Simplify and unify file structure
