+++ server/src/requests/music-analysis/player.js
@@ -1,4 +1,4 @@
-const axiosInstance = require("../../../utils/customAxios").axiosSpotifyAPI;
+const axiosInstance = require("../../utils/customAxios").axiosSpotifyAPI;
 
 module.exports.getPlayed = async (authCodeToken, limit) => {
   try {
